refactor(coyote): cleanse overall package structure
