+++ tomcat/src/main/java/org/apache/coyote/request/HttpRequest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.support;
+package org.apache.coyote.request;
 
 import java.util.HashMap;
 import java.util.List;
@@ -6,26 +6,27 @@ import java.util.Map;
 import java.util.Optional;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import org.apache.coyote.support.HttpMethod;
 
 public class HttpRequest {
 
-    private final Method method;
+    private final HttpMethod method;
     private final String uri;
     private final Map<String, String> parameters;
 
-    private HttpRequest(Method method, String uri, Map<String, String> parameters) {
+    private HttpRequest(HttpMethod method, String uri, Map<String, String> parameters) {
         this.method = method;
         this.uri = uri;
         this.parameters = parameters;
     }
 
-    private HttpRequest(Method method, String uri) {
+    private HttpRequest(HttpMethod method, String uri) {
         this(method, uri, new HashMap<>());
     }
 
     public static HttpRequest of(List<String> request) {
         final var startLine = request.get(0).split(" ");
-        final var method = Method.valueOf(startLine[0]);
+        final var method = HttpMethod.valueOf(startLine[0]);
         final var uri = startLine[1];
         if (uri.contains("?")) {
             final var delimiterIndex = uri.indexOf("?");
@@ -51,7 +52,7 @@ public class HttpRequest {
     }
 
     public boolean isGet() {
-        return this.method.equals(Method.GET);
+        return this.method.equals(HttpMethod.GET);
     }
 
     public Optional<User> checkLoginAccount() {
