+++ tomcat/src/main/java/org/apache/coyote/response/ResourceView.java
@@ -1,10 +1,11 @@
-package org.apache.coyote.support;
+package org.apache.coyote.response;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import org.apache.exception.HttpException;
+import org.apache.coyote.support.HttpStatus;
+import org.apache.coyote.exception.HttpException;
 
 public class ResourceView {
