+++ tomcat/src/main/java/org/apache/coyote/support/HttpMethod.java
@@ -1,5 +1,5 @@
 package org.apache.coyote.support;
 
-public enum Method {
+public enum HttpMethod {
     GET, POST, PUT, DELETE
 }
